% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{add_new_dimension_levels_full}
\alias{add_new_dimension_levels_full}
\title{Add new dimension levels to existing table}
\usage{
add_new_dimension_levels_full(code, level, con, schema = "platform")
}
\arguments{
\item{code}{surs table code such as 0457102S}

\item{level}{level code such as 29}

\item{con}{connection to database}

\item{schema}{database schema name}
}
\value{
nothing
}
\description{
Sometimes an existing table will get a new category. In that case you want to
add the levels, series and series levels separately just for the new series
and this function does that. You need to get the code for the new level from
Si-Stat
}
