% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{add_new_table}
\alias{add_new_table}
\title{Umbrella code for adding a new table to the database}
\usage{
add_new_table(code, con)
}
\arguments{
\item{code}{character ID of the table e.g. "0714621S"}

\item{con}{connection to database}
}
\value{
nothing
}
\description{
The code gets the new hierarchy from the structAPI to get the category levels
right and then inserts all the required strucutres and finally the data points
for the first set of vintages for these series
}
