% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_meritve_level_units}
\alias{add_meritve_level_units}
\alias{add_valuenotes_level_units}
\title{Joining the unit tables from either meritve or valuenotes.}
\usage{
add_meritve_level_units(
  expanded_level_codes,
  meritve_dim_no,
  units_by_meritve_levels
)

add_valuenotes_level_units(
  expanded_level_codes,
  valuenotes_dim_no,
  units_by_levels
)
}
\arguments{
\item{expanded_level_codes}{dataframe output of \link[SURSfetchR]{expand_to_level_codes}}

\item{meritve_dim_no}{numeric output of \link[SURSfetchR]{get_meritve_no}}

\item{units_by_meritve_levels}{dataframe output of \link[SURSfetchR]{get_unit_levels_from_meritve}}

\item{valuenotes_dim_no}{numeric output of \link[SURSfetchR]{get_valuenotes_no}}

\item{units_by_levels}{dataframe output of \link[SURSfetchR]{get_unit_levels_from_meritve}}
}
\value{
datafram from expanded_level_codes with added `unit_id` column
}
\description{
These two helper functions take the expanded level code tables and
join the appropriate unit_ids from either the meritve or the valuenotes
tables.
}
\keyword{internal}
