% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_setup_n_utilities.R
\name{build_db_tables}
\alias{build_db_tables}
\title{Build all database tables}
\usage{
build_db_tables(con, schema = "test_platform")
}
\arguments{
\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{schema}{character schema name, default is "test_platform"}
}
\description{
Creates all the tables required to run the database in a given schema by
running the appropriate sql file. (Excluded from testing).
}
