% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{email_surs_changes}
\alias{email_surs_changes}
\title{Email nicely parsed SURS reported changes}
\usage{
email_surs_changes(daily, recipient = "maja.zaloznik@gmail.com")
}
\arguments{
\item{daily}{table output of \link[SURSfetchR]{extract_new_changes}}

\item{recipient}{single email of recipient}
}
\value{
null - side effect is sending an email.
}
\description{
This email takes the outputs of \link[SURSfetchR]{extract_new_changes} that have
been formatted with the helper function \link[SURSfetchR]{print_for_type} and
emails it from a hardcoded gmail account - currently mine, but I'll change
it to an Umar one. Not included in testing suite.
}
