% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_setup_n_utilities.R
\name{execute_sql_file}
\alias{execute_sql_file}
\title{Generic run sql file}
\usage{
execute_sql_file(con, file, schema = "platform")
}
\arguments{
\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{file}{path to sql file}

\item{schema}{character schema name, default is "platform"}
}
\description{
Executes an sql file. Uses dbExecute, which requires single statements to
be passed i.e. cannot handle a whole file, so the function splits the file
on `;`. Potentially not safe.
If file has stored procedures, use \link[SURSfetchR]{execute_sql_functions_file}
instead.
}
\details{
Inspiration from `timeseriesdb` package by Matt Bannert.
}
