% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{extract_todays_changes}
\alias{extract_todays_changes}
\title{Extract changes that are due today}
\usage{
extract_todays_changes(new_df, date = Sys.Date())
}
\arguments{
\item{new_df}{dataframe output from \link[SURSfetchR]{surs_change_api}}

\item{date}{desired single date, defaults to today.}
}
\value{
dataframe with 7 columns (no publication date, because we don't need here)
}
\description{
Get dataframe of changes that are coming into effect on today's date - or
another specified date
}
