% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelz.R
\name{full_hierarchy_unnest}
\alias{full_hierarchy_unnest}
\title{Extract time dim stuff from nested tables in fully joined matrix & level table}
\usage{
full_hierarchy_unnest(df)
}
\arguments{
\item{df}{dataframe output of \link[SURSfetchR]{matrix_n_level_join}}
}
\value{
dataframe with three more columns than going in: `dim_names_notime` are
the dimensions without the time dimension, `dim_lz_notime` are the numbers of
levels of these dimensions and `no_series` is the number of series in the matrix
}
\description{
With the fully joined table output from \link[SURSfetchR]{matrix_n_level_join},
this function extracts some more stuff from the nested listcolumns especially
the dimensions that are not time related and calculates the number of
series in each table (`no_series`).T
}
\details{
This is not foolproof in the sense that the time dimension must be explicitly
flagged with the time argument in the header, so if SURS doesn't do that,
then the matrix won't show up here
}
