% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructure_API.R
\name{get_API_response}
\alias{get_API_response}
\title{Get API response in json format and extract as text}
\usage{
get_API_response(
  url = "https://pxweb.stat.si/SiStat/sl/Api/GetStructure",
  ua = httr::user_agent("https://github.com/majazaloznik/SURSfetchR")
)
}
\arguments{
\item{url}{in case you want to call some other API, but defaults to the GetStructure one}

\item{ua}{thought i'd be nice and identify myself in the call, the user agent
gets put in the call header in case it might be helpful.}
}
\value{
(large) character vector with json response from the API to be parsed and handled
downstream
}
\description{
This function is a general funciton to return a simple json response in text
format. The defaults are set to return from the GetStructure method on the
SiStat API with the complete file (matrix) and category (field) structure
for the available .px files, but can be used for anything. Added the user agent
as well, currenlty this package repository.
}
