% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructure_API.R
\name{get_full_category_hierarchy}
\alias{get_full_category_hierarchy}
\title{Get full category hierarchy from the API}
\usage{
get_full_category_hierarchy()
}
\value{
dataframe with flattened tree hierarchy
}
\description{
Wrapper function
}
