% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_full_structure}
\alias{get_full_structure}
\title{Extract proper hierarcy from tree structure}
\usage{
get_full_structure(tree)
}
\arguments{
\item{tree}{data.tree structure from \link[SURSfetchR]{parse_structAPI_response}}
}
\value{
a data frame with 6 columns and a row for each matrix and each field.
}
\description{
The tree structure has superfluous levels that need to be collapsed, which this function
does, while pulling out the relevant labels and other fields we are interested in and
outputs a data frame.
}
