% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reading.R
\name{get_interval_id}
\alias{get_interval_id}
\title{Helper to get interval id from lookup vector}
\usage{
get_interval_id(interval_text)
}
\arguments{
\item{interval_text}{text of the time interval name}
}
\value{
character code of interval id
}
\description{
This is not a db reading function, but it kinda fits with them, simply
because i'm using a lookup vector inside the function instead of creating
another table in the database to get the interval id.
}
\keyword{internal}
