% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelz.R
\name{get_level_combos}
\alias{get_level_combos}
\title{Helper function for extracting level combinations for each matrix}
\usage{
get_level_combos(x, y)
}
\arguments{
\item{x}{name of listcolumn with levelz tibbles}

\item{y}{name of listcolumn with dimension names to be used (so you can use
a subset - usually the subset excluding the time variable)}
}
\value{
a dataframe or tibble with a column for each level and a row for each
combination of level values.
}
\description{
Starting from the dataframe output from \link[SURSfetchR]{full_hierarchy_unnest}
this helper function gets the non-time levels from the `levelz` table, specifically
their names and creates a table with all possible combinations i.e. a table listing
all the series inside each matrix. Can be used `rowwise` in a `mutate` call with
something like  `list(get_level_combos(levelz, dim_names_notime)))`, but don't
try running it on the whole table, since it would end up with 415 million rows
and that won't work..
}
