% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructure_API.R
\name{get_matrix_hierarchy}
\alias{get_matrix_hierarchy}
\title{Extract only the matrix hierarchy from the full one}
\usage{
get_matrix_hierarchy(full_df)
}
\arguments{
\item{full_df}{data frame output from  \link[SURSfetchR]{get_full_structure}}
}
\value{
a data frame with 7 columns and a row for each matrix
}
\description{
Extracts only rows describing matrices from the full hierarchy.
}
