% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_px.R
\name{get_parent}
\alias{get_parent}
\title{Helper function to get parent category from full hierarchy}
\usage{
get_parent(id_no, full, row, output)
}
\arguments{
\item{id_no}{id of category looking for its parent}

\item{full}{full hierarchy of matrixes and fields out of
\link[SURSfetchR]{get_full_structure}.}

\item{row}{subset of full with relevant ids passed from \link[SURSfetchR]{get_row}}

\item{output}{output table}
}
\value{
dataframe with four columns with id, name, parent id and source id (which is
1 for SURS)
}
\description{
This is a recursive function to be used inside another semi-recursive
function \link[SURSfetchR]{get_row} to extract the ids of categories and
their parents.
}
