% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_px.R
\name{get_px_data}
\alias{get_px_data}
\title{Get the data for an individual table}
\usage{
get_px_data(id)
}
\arguments{
\item{id}{character vector of length 1 with code of matrix. Can be with or
without the .px extension.}
}
\value{
A long data.frame with all the data
}
\description{
Get the full data table for each SURS table (matrix).
}
