% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_row}
\alias{get_row}
\title{Semi-recursive function to get out the category hierarchy}
\usage{
get_row(id_no, full, output = NULL)
}
\arguments{
\item{id_no}{id of category looking for its parent}

\item{full}{full hierarchy of matrixes and fields out of
\link[SURSfetchR]{get_full_structure}.}

\item{output}{output table}
}
\value{
dataframe with four columns with id, name, parent id and source id (which is
1 for SURS), with unique rows, ready to be inserted into the database.
}
\description{
Using the recursive \link[SURSfetchR]{get_parent} in a weird way that works
in most cases - but doensn't work that great when a category has two parents,
which on occasion happens. Anyway, this is solved using distinct, so the
results are fine, just the internals are not as efficient as they could be
if i was a bit smarter.
}
