% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reading.R
\name{get_meritve_no}
\alias{get_meritve_no}
\alias{get_unit_levels_from_meritve}
\title{Helper functions to get stuff out of db tables}
\usage{
get_meritve_no(tbl_id, con, schema = "platform")

get_unit_levels_from_meritve(meritve_level_text, con, schema = "platform")
}
\arguments{
\item{tbl_id}{numeric table id from table table}

\item{con}{connection to postgres database}

\item{code_no}{px. code e.g. 0300230S}

\item{unit}{char text of unit label}

\item{dim_name}{character dimension name}

\item{tbl_dm_id}{numeric table-dimension id from that same table}

\item{lvl_text}{character text of level label}

\item{meritve_dim_id}{numeric dimension id of MERITVE dim}
}
\value{
numeric position of MERITVE dimension if it exists,

tibble with 4 cols including `level_value` and `unit_id`
}
\description{
Group of functions that lookup stuff in relevant tables on the `con` connection
to get correct table id from px code. Also make sure the `search_path` is set to
the correct scheme.
}
\details{
NB: there is an issue with BIGINT datatype being read by dittodb in R, that's
why those have an `as.numeric()` added at the end, because there is no native
type in R or sth.
}
\keyword{internal}
