% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelz.R
\name{get_surs_metadata}
\alias{get_surs_metadata}
\title{GET the metadata for an individual table from the SURS API}
\usage{
get_surs_metadata(id)
}
\arguments{
\item{id}{character vector of length 1 with code of matrix. Can be with or
without the .px extension.}
}
\value{
A tibble with four columns and the same number of rows as there
are dimensions in the table
}
\description{
GET the metadata for an individual table from the SURS API
}
