% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{get_table_id}
\alias{get_table_id}
\title{Helper fun to get correct table id from px code}
\usage{
get_table_id(code_no)
}
\arguments{
\item{code_no}{px. code e.g. 0300230S}
}
\value{
numeric code from the table table
}
\description{
Helper fun to get correct table id from px code
}
\keyword{internal}
