% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelz.R
\name{get_table_levels}
\alias{get_table_levels}
\title{GET the dimensions and levels for an individual table from the SURS API}
\usage{
get_table_levels(id)
}
\arguments{
\item{id}{character vector of length 1 with code of matrix. Can be with or
without the .px extension.}
}
\value{
A tibble with four columns and the same number of rows as there
are dimensions in the table
}
\description{
Takes the matrix code of a SURS table (with or without the .px extension)
and uses GET to return the metadata as a tibble. NB: this is a limited set
of metadata, with only dim_names and levels and the elim and time varz.
I think I used it because it had the levels in a nice format. Use XXX instead
to get the richer metadata out.
}
