% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_tree}
\alias{get_tree}
\title{Get tree out of hierarchical df}
\usage{
get_tree(df)
}
\arguments{
\item{df}{data frame output from \link[SURSfetchR]{get_field_hierarchy} or
\link[SURSfetchR]{get_matrix_hierarchy}}
}
\value{
a data.tree tree object
}
\description{
Takes the df output from \link[SURSfetchR]{get_field_hierarchy} or
\link[SURSfetchR]{get_matrix_hierarchy} and creates a data.tree
}
