% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{get_unit_id}
\alias{get_unit_id}
\title{Helper fun to get correct code from unit text}
\usage{
get_unit_id(unit)
}
\arguments{
\item{unit}{char text of unit label}
}
\value{
numeric code from db table unit
}
\description{
Helper fun to get correct code from unit text
}
\keyword{internal}
