% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{Id2}
\alias{Id2}
\title{Wrapper for database identifier construction}
\usage{
Id2(table, schema = "test_platform")
}
\arguments{
\item{table}{character name of table}

\item{schema}{character name of schema, default currently `test_platform`}
}
\value{
Id object
}
\description{
Wraps `Id` to a default schema and without requiring named args.
}
