% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_data_points}
\alias{insert_data_points}
\title{Insert datapoints into data_point table}
\usage{
insert_data_points(code_no, con)
}
\arguments{
\item{code_no}{character SURS px. code e.g. 0300230S}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
nothing, just some printing along the way
}
\description{
This is one hell of a function. Not even sure how i got it all to work.. It
should eventually get rewritten like all the other insert functions are in
PL/pgSQL, but if it ain't broke don't try to fix it, innit?
}
\details{
So, the function downloads and preps the data with \link[SURSfetchR]{prepare_data_table}
and writes it to a temporary table in the database. Then it gets the table id,
dimension ids and dim names. Then with all that in hand, the main bit of the code
prepares the temp table by adding the appropriate vintage ids for each series.

Then the function gets the "za?asni podatki", which are inside the fucking time
period variable.. Needs to strip them off and save them separately, because we do
want to know.

Then to finish off the function inserts any new periods into the period table,
adds the data points to the data point table and the flags (temporary = T) into
the flag-datapoint table.
}
