% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_table_structures}
\alias{insert_new_table_structures}
\title{Insert table structure data for a new table}
\usage{
insert_new_table_structures(code_no, con, full)
}
\arguments{
\item{code_no}{character SURS px. code e.g. 0300230S}

\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{full}{full field hierarchy with parent_ids et al, output from
\link[SURSfetchR]{get_full_structure}}
}
\value{
list of tables with counts for each inserted row.
}
\description{
When a new table (in SURS speak 'matrix") is added, a set of nine
tables need to be populated with appropriate metadata about that table.
This umbrella function calls the respective SQL functions for each
of the nine tables.
}
\examples{
\dontrun{
purrr::walk(master_list_surs$code, ~insert_new_table_structures(.x, con, full))
}
}
