% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_script_run}
\alias{log_script_run}
\title{Log script run}
\usage{
log_script_run(
  path = here::here("logs"),
  script = envDocument::getScriptPath()
)
}
\arguments{
\item{path}{path to logs folder, defaults to `here::here("logs")`}

\item{script}{path of script being run, defaults to actual script being run,
which only works when beign sources.}
}
\value{
NA if there is an error, otherwise nothing, but has side effects
of writing to log files.
}
\description{
Simple helper function that creates a log if it doesn't exist yet in the `logs`
folder and appends a row logging a script run. Logs in two logs, one for
the individual script file and one common log for all the scripts, which
also includes the script name. Logs are dataframes saved as .rds files.
}
