% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructure_API.R
\name{node_file_id}
\alias{node_file_id}
\title{Helper - get px file ids}
\usage{
node_file_id(node)
}
\arguments{
\item{node}{tree node}
}
\value{
character vector of length 1
}
\description{
Getting data from the depths of the API response. Helper function used
to reshape and extract the data structure. Gets the ID of the matrix file
(which is different from the matrix ID, since it changes with every update).
}
\keyword{internal}
