% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{node_id}
\alias{node_id}
\title{Helper - get node ids}
\usage{
node_id(node)
}
\arguments{
\item{node}{tree node}
}
\value{
character vector of length 1
}
\description{
Getting data from the depths of the API response. Helper function used
to reshape and extract the data structure. Gets the ID of the "podrocje"
or of the px matrix.
}
\keyword{internal}
