% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructure_API.R
\name{node_matrix_name}
\alias{node_matrix_name}
\title{Helper - get matrix (table) name}
\usage{
node_matrix_name(node)
}
\arguments{
\item{node}{tree node}
}
\value{
character vector of length 1
}
\description{
Getting data from the depths of the API response. Helper function used
to reshape and extract the data structure. Gets the long form name of the
matrix.
}
\keyword{internal}
