% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{node_name}
\alias{node_name}
\title{Helper - get node names}
\usage{
node_name(node)
}
\arguments{
\item{node}{tree node}
}
\value{
character vector of length 1
}
\description{
Getting data from the depths of the API response. Helper function used
to reshape and extract the data structure. Gets the name of the "podrocje"
or the px matrix code
}
