% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructure_API.R
\name{node_parent}
\alias{node_parent}
\title{Helper - get matrix (table) parent}
\usage{
node_parent(node)
}
\arguments{
\item{node}{tree node}
}
\value{
character vector of length 1
}
\description{
Getting data from the depths of the API response. Helper function used
to reshape and extract the data structure. Gets the node's parent node name.
Of course if the parent name is (archive)matrixList, it get's the next parent.
}
\keyword{internal}
