% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{node_updated}
\alias{node_updated}
\title{Helper - get matrix (table) last update time}
\usage{
node_updated(node)
}
\arguments{
\item{node}{tree node}
}
\value{
character vector of length 1
}
\description{
Getting data from the depths of the API response. Helper function used
to reshape and extract the data structure. Gets the last updated time
for a matrix
}
