% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{parse_row}
\alias{parse_row}
\title{Parse single row of changes for printing in email}
\usage{
parse_row(row)
}
\arguments{
\item{row}{single row from table output of \link[SURSfetchR]{extract_new_changes}}
}
\value{
a character string in html
}
\description{
This function takes a single 7 column row from the daily changes table output
from \link[SURSfetchR]{extract_new_changes} and parses into a nice printable
format that can then be used in a reporting email.
}
