% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{parse_structAPI_response}
\alias{parse_structAPI_response}
\title{Parse API response into usable format}
\usage{
parse_structAPI_response(res, output = "tree")
}
\arguments{
\item{res}{API text response - output from \link[SURSfetchR]{get_structAPI_response}}

\item{output}{defaults to "tree", other possibilities are "lol" for a list of lists
and "df" for a df obvs.}
}
\value{
a data.tree tree object, a data frame or a list of lists
}
\description{
Defaults to returning a tree structure from the package data.tree, but
you could also return a data frame or a list of lists.
}
