% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{parse_surs_updates}
\alias{parse_surs_updates}
\title{Get the codes of updated tables}
\usage{
parse_surs_updates()
}
\value{
a dataframe with two columns and over 3700 rows with the id of the
table and the date it was last updated.
}
\description{
This function parses the html list on the SURS website which seems to be the
only outside facing endpoint where you can get access to the update date
for each of their published tables. And no, the API doesn't have this info.
}
\details{
[This](https://pxweb.stat.si/SiStatData/pxweb/sl/Data/) is the table and we
want the date from the alt text on the little clock icon (the date of the
most recent change) and the id of the table, which is in the link.
}
