% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_category_table}
\alias{prepare_category_table}
\title{Prepare table to insert into `category` table}
\usage{
prepare_category_table(code_no, full)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{full}{full field hierarchy with parent_ids et al, output from
\link[SURSfetchR]{get_full_structure}}
}
\value{
a dataframe with the `id`, `name`, `source_id` for each category that
the table is a member of.
}
\description{
Helper function that extracts all the parent categories from the full
hierarchy data.frame, and prepares the category table with field ids and
their names. Returns table ready to insert into the `category` table with the db_writing family
of functions.
}
