% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_category_table_table}
\alias{prepare_category_table_table}
\title{Prepare table to insert into `category_table` table}
\usage{
prepare_category_table_table(code_no, full, con)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{full}{full field hierarchy with parent_ids et al, output from
\link[SURSfetchR]{get_full_structure}}
}
\value{
a dataframe with the `code`, `name`, `source`, `url`, and `notes` columns
for this table.
}
\description{
Helper function that extracts the parent category for each table from the full
hierarchy data.frame, and fills up the category_table table with the table ids and
their categories (parents)
}
