% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data_table}
\alias{prepare_data_table}
\title{Get and prepare data for import}
\usage{
prepare_data_table(code_no, con, schema)
}
\arguments{
\item{code_no}{character SURS px. code e.g. 0300230S}

\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{schema}{character schema name, default is "platform"}
}
\value{
a dataframe that was downloaded from the .px file, but
with recoded dimension levels and filtered series we arent' tracking.
}
\description{
Downloads and prepares the timeseries data for importing into the database.
Because the original data format has the full labels of the levels, these
are recoded into the alphanumeric codes.
The datapoints are filtered to only include the series that are in the
database in case some levels have been removed during the structure import.
}
