% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_dimension_levels_table}
\alias{prepare_dimension_levels_table}
\title{Prepare table to insert into `dimension_levels` table}
\usage{
prepare_dimension_levels_table(code_no, con)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{con}{connection to the database
\link[SURSfetchR]{get_full_structure}}
}
\value{
a dataframe with the `code`, `name`, `source`, `url`, and `notes` columns
for this table.
}
\description{
Helper function that extracts the levels for all the dimensions for each
table and get their code and text.
}
