% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_series_levels_table}
\alias{prepare_series_levels_table}
\title{Prepare table to insert into `series_levels` table}
\usage{
prepare_series_levels_table(code_no, con)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{con}{connection to the database}
}
\value{
a dataframe with the `code`, `name`, `source`, `url`, and `notes` columns
for this table.
}
\description{
Helper function that extracts the individual levels for each series and
gets the correct dimension id for each one and the correct series id to
keep with the constraints.
}
