% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_series_table}
\alias{prepare_series_table}
\title{Prepare table to insert into `series` table}
\usage{
prepare_series_table(code_no, con)
}
\arguments{
\item{code_no}{character obect of the matrix code (e.g. 2300123S)}

\item{con}{postgresql connection to the database.}
}
\value{
a dataframe with three columns: series_title, series_code and unit_id as
well as the same number of rows as there are series
}
\description{
This is a big one. Prepares the table to insert into the series table, which
along expanding the leves to get all the series and their codes as well, which
include also figuring out their time interval, this function also tries to
extract the unit for each series, which can get pretty tricky. There are generally
three options:
1. either the whole table and therefore all the series within it have a single unit
which we get from the .px metadata.
2. or there is a MERITVE dimensions, which might (?) contain the units for each
level in parentheses at the end of the label. So we get if from there
3. ooor, there is no MERITVE dimension, and then there is (?) a slot in the .px
data called `valuenotes`, which contains the units and again there is an even
more convoluted way to get them out and know which level values they apply to.
}
