% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_table_dimensions_table}
\alias{prepare_table_dimensions_table}
\title{Prepare table to insert into `table_dimensions` table}
\usage{
prepare_table_dimensions_table(code_no, con)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{con}{connection to the database}
}
\value{
a dataframe with the `code`, `name`, `source`, `url`, and `notes` columns
for this table.
}
\description{
Helper function that extracts the dimensions for each table and their "time" status.
}
