% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_units_table}
\alias{prepare_units_table}
\title{Prepare table to insert into `units` table}
\usage{
prepare_units_table(code_no, con)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{con}{connection to the database}
}
\value{
a dataframe with the `code`, `name`, `source`, `url`, and `notes` columns
for this table.
}
\description{
Helper function that extracts the units for each table from the px metadata.
}
