% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{print_for_type}
\alias{print_for_type}
\title{Prints out changes for a single change type}
\usage{
print_for_type(daily, type = 2)
}
\arguments{
\item{daily}{table output of \link[SURSfetchR]{extract_new_changes}}

\item{type}{number from 1 to 6 for each different type of reported change.}
}
\value{
a character string in html
}
\description{
This function takes as its input the daily changes table output
from \link[SURSfetchR]{extract_new_changes} and for each type of change (1-6)
prints out an emailable text.
}
