% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{recode_labels}
\alias{recode_labels}
\title{Helper for recoding dimension levels}
\usage{
recode_labels(i, codes, labels, df, time_dim)
}
\arguments{
\item{i}{numeric index for mapping over the dimensions}
}
\value{
df with the ith dimension recoded from labels to codes.
}
\description{
Recodes dimension level labels into dimension level codes. Both
labels and codes are extracted inside the  \link[SURSfetchR]{prepare_data_table}
function, which is where this helper is also called.
}
\keyword{internal}
