% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_setup_n_utilities.R
\name{sql_function_call}
\alias{sql_function_call}
\title{Construct and execute an SQL function call}
\usage{
sql_function_call(con, fun_name, args, schema = "test_platform")
}
\arguments{
\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{args}{}

\item{schema}{character schema name, default is "test_platform"}
}
\description{
Constructs the call for the funciton `schema`.`fun_name` fro the database
with the arguments `args` and returns the result.
}
\details{
Inspiration from `timeseriesdb` package by Matt Bannert.
}
