% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{subset_parsed_df}
\alias{subset_parsed_df}
\title{Subset parsed df by date updated}
\usage{
subset_parsed_df(df, date = Sys.Date())
}
\arguments{
\item{df}{output of \link[SURSfetchR]{parse_surs_updates}}

\item{date}{date to subset the table by, if empty defaults to today's date,
if NULL returns whole table.}
}
\value{
table dataframe with two columns, nrow depends on subset.
}
\description{
Subset parsed df by date updated
}
