% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{surs_change_api}
\alias{surs_change_api}
\title{Get complete list of published changes}
\usage{
surs_change_api(body = NULL)
}
\arguments{
\item{body}{body to pass to the GET request, but really only required for
testing, because there is no parameters required on this request.}
}
\value{
a data frame with 7 columns
}
\description{
SURS set up another convenience API with the data used for
[this list](https://pxweb.stat.si/SiStat/en/Notifications) where they publish
all the anticipated changes to the database structure. This funciton pulls the
data and parses it into a neat dataframe with only relevant (slovenian) columns.
}
