% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{SURS_import_data_points}
\alias{SURS_import_data_points}
\title{Insert data points from SURS}
\usage{
SURS_import_data_points(code_no, con, schema = "platform")
}
\arguments{
\item{code_no}{SURS code name of the table}

\item{con}{Database connection}

\item{schema}{Schema name}
}
\value{
Insertion results (invisibly)
}
\description{
Function to prepare and insert SURS data points. The function first prepares
the required vintages and inserts them, then prepares the data points
table and isnerts it. The function returns the results invisibly.
}
\details{
This is a SURS specific function, which should be followed by the generic
UMARimportR function to write the vintage hashes and clean up redundant
vintages.
}
