% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{SURS_insert_data_points}
\alias{SURS_insert_data_points}
\title{Insert data points from SURS}
\usage{
SURS_insert_data_points(code_no, con, schema = "platform")
}
\arguments{
\item{code_no}{SURS code name of the table}

\item{con}{Database connection}

\item{schema}{Schema name}
}
\value{
Insertion results (invisibly)
}
\description{
Function to prepare and insert SURS data points.
}
