% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{surs_opsi_api}
\alias{surs_opsi_api}
\title{Get daily table changes from SURS' OPSI API}
\usage{
surs_opsi_api(date = Sys.Date(), table = "Update")
}
\arguments{
\item{date}{required, defaults to today.}

\item{table}{One of "New", "Update" or "Delete", defaulting to "Update"}
}
\value{
a dataframe with three columns, or an error message if there was no data.
}
\description{
SURS set up a convenience mini API for OPSI so they could get info on the
daily changes. It returns New, Updated and Deleted tables. This function
gets the response from the API for a particular date and desired table and
cleans it up to keep just the date, clean table id code and title (the latter
is just for kicks).
}
