% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURSfetchR-package.R
\docType{package}
\name{SURSfetchR-package}
\alias{SURSfetchR}
\alias{SURSfetchR-package}
\title{SURSfetchR: Fetching data from the SURS API}
\description{
Checking for tables on the SURS API that have been updated (or created or deleted), downloading the data, and preparing it for the UMAR data model.
}
\author{
\strong{Maintainer}: Maja Zaloznik \email{maja.zaloznik@gmail.com}

}
\keyword{internal}
