% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{update_change_table}
\alias{update_change_table}
\title{Append new changes to change table}
\usage{
update_change_table(old_df, changes)
}
\arguments{
\item{old_df}{dataframe output from \link[SURSfetchR]{update_change_table}}

\item{changes}{dataframe output from \link[SURSfetchR]{extract_new_changes}}
}
\value{
dataframe with 8 columns
}
\description{
Take the new changes extracted from the API by comparing with the change table
add today's date and append to previous change table
}
