% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{write_multiple_rows}
\alias{write_multiple_rows}
\title{Umbrella function to write multiple rows into the a postgres table}
\usage{
write_multiple_rows(master_list_surs, con, table_name, sql_statement)
}
\arguments{
\item{master_list_surs}{dataframe with at minimum a "code" column with
the SURS matrix code for each table to be inserted.}

\item{con}{a connection to a postgres database with an approproate `table`
table.}

\item{table_name}{character name of table in db, also used to get the
relevant function to write a single row.}

\item{sql_statement}{character string to write single row to the database.}
}
\value{
nothing - side effect is writing to the database, and outputing message
to say how many rows were inserted.
}
\description{
Takes a column of SURS codes from the master_list_code and for each row
gets the metadata and writes it to the appropriate table  using the
passed sql statement and the appropriate function from the `write_row`
family of functions for each single row.
}
