% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{write_row_category_relationship}
\alias{write_row_category_relationship}
\title{Write categories for a single table to the `category_relationship` table}
\usage{
write_row_category_relationship(code_no, con, sql_statement, counter, full)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{con}{connection to the database}

\item{sql_statement}{the sql statement to insert the values}

\item{counter}{integer counter used in  \link[SURSfetchR]{write_multiple_rows}
to count how many successful rows were inserted.}

\item{full}{full field hierarchy with parent_ids et al, output from
\link[SURSfetchR]{get_full_structure}}
}
\value{
incremented counter, side effect is writing to the database.
}
\description{
Helper function that extracts the field hierarchy from the full
hierarchy data.frame, and fills up the category table with field ids and
their parents. Gets run from \link[SURSfetchR]{write_multiple_rows}.
}
