% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{write_row_series_levels}
\alias{write_row_series_levels}
\title{Write series-level combinations into to the `series_levels` table}
\usage{
write_row_series_levels(code_no, con, sql_statement, counter, ...)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{con}{connection to the database}

\item{sql_statement}{the sql statement to insert the values}

\item{counter}{integer counter used in  \link[SURSfetchR]{write_multiple_rows}
to count how many successful rows were inserted.}

\item{...}{just here, because other funs in this family have extra parameters
passed to them and i cannot use map unless this one also has this option.}
}
\value{
incremented counter, side effect is writing to the database.
}
\description{
Helper function that extracts the individual levels for each series and
gets the correct dimension id for each one and the correct series id to
keep with the constraints.
}
\details{
Gets run from \link[SURSfetchR]{write_multiple_rows}.
}
