% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{write_row_table}
\alias{write_row_table}
\title{Write a single row to the `table` table for SURS}
\usage{
write_row_table(code_no, dbtable, con, sql_statement, counter)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{dbtable}{the `tbl()` link to the table we are inserting to}

\item{con}{connection to the database}

\item{sql_statement}{the sql statement to insert the values}

\item{counter}{integer counter used in  \link[SURSfetchR]{write_multiple_rows}
to count how many successful rows were inserted.}
}
\value{
side effect is writing to the database.
}
\description{
Helper function that extracts metadata from the API and inserts it into the
sql statement and then gets written to the `table` table in the connection.
}
\details{
Checks to make sure the code_no doesn't exist in the table already.
}
