% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{write_row_unit}
\alias{write_row_unit}
\title{Write units into to the `units` table}
\usage{
write_row_unit(code_no, dbunits, con, sql_statement, counter, ...)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{dbunits}{tbl() link to db table, although this one isn't actually used}

\item{con}{connection to the database}

\item{sql_statement}{the sql statement to insert the values}

\item{counter}{integer counter used in  \link[SURSfetchR]{write_multiple_rows}
to count how many successful rows were inserted.}

\item{...}{just here, because other funs in this family have extra parameters
passed to them and i cannot use map unless this one also has this option.}
}
\value{
incremented counter, side effect is writing to the database.
}
\description{
Helper function that extracts the units for each table from the px metadata.
Gets run from \link[SURSfetchR]{write_multiple_rows}.
}
