% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{write_rows_table}
\alias{write_rows_table}
\title{Umbrella function to write multiple rows into the `table` table}
\usage{
write_rows_table(master_list_surs, con)
}
\arguments{
\item{master_list_surs}{dataframe with at minimum a "code" column with
the SURS matrix code for each table to be inserted.}

\item{con}{a connection to a postgres database with an approproate `table`
table.}
}
\value{
nothing - side effect is writing to the database.
}
\description{
Takes a column of SURS codes from the master_list_code and for each row
gets the metadata and writes it to the `table` table  using the
\link[SURSfetchR]{write_row_table} function for each single row.
}
