% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_setup.R
\name{build_db_tables}
\alias{build_db_tables}
\title{Build all database tables}
\usage{
build_db_tables(con, schema = "test_platform")
}
\arguments{
\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{schema}{character schema name, defaults to 'timeseries'.}

\item{prnt}{function log printing function}
}
\description{
Creats all the tables required to run the database in a given schema.
Inspiration from `timeseriesdb` package.
}
