% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{email_surs_changes}
\alias{email_surs_changes}
\title{Email nicely parsed SURS reported changes}
\usage{
email_surs_changes(body, recipient = "maja.zaloznik@gmail.com")
}
\arguments{
\item{body}{html character string output of  \link[SURSfetchR]{email_surs_changes_body}}

\item{recipient}{single email of recipient}
}
\value{
null - side effect is sending an email.
}
\description{
This function takes a html text as the body from  \link[SURSfetchR]{email_surs_changes_body}
and emails it from a hardcoded gmail account - the Umar Data Bot one one.
Not included in testing suite. This will only work with the OAuth secrets in
the gm_autn() setup from wherever this function is run.
}
