% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{expand_to_level_codes}
\alias{expand_to_level_codes}
\alias{expand_to_series_titles}
\title{Expanding from levels to series codes and titles}
\usage{
expand_to_level_codes(tbl_id, con, schema = "platform")

expand_to_series_titles(tbl_id, con, schema = "platform")
}
\arguments{
\item{code_no}{code e.g. 0300230S}
}
\value{
dataframe with expanded levels, one column per non-time dimension plus
unit_id for the level codes and sinle column with series titles for the other one.
}
\description{
These two helper functions take a set of non-time levels for a single table
and expand the grid to get all of their combinations and then either return
a dataframe with columns for each level code, or one where the level texts
have been concatenated into the series titles.
}
\keyword{internal}
