% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{extract_new_changes}
\alias{extract_new_changes}
\title{Compare newly parsed table with old table to see changes in Notifications API}
\usage{
extract_new_changes(new_df, old_df)
}
\arguments{
\item{new_df}{dataframe output from \link[SURSfetchR]{surs_change_api}}

\item{old_df}{dataframe output from \link[SURSfetchR]{update_change_table}}
}
\value{
dataframe with 7 columns
}
\description{
A function to extract the newly published changes on the Notificaitons API since
the script was previously run.
}
