% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelz.R
\name{fill_listcolumn_w_mtdt}
\alias{fill_listcolumn_w_mtdt}
\title{GET metadata and join into dataframe of matrices}
\usage{
fill_listcolumn_w_mtdt(df)
}
\arguments{
\item{df}{dataframe with an id column containing the .px codes}
}
\value{
a dataframe appended with the listcolumn of metadata
}
\description{
Starting out with a data frame with an `id` column of matrix names, the
function loops through them and parses the GET response metadata into
a tibble that is saved in the `levelz` list-column. Uses
\link[SURSfetchR]{get_surs_metadata} in the loop.
}
