% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_px.R
\name{get_single_unit_from_px}
\alias{get_single_unit_from_px}
\alias{get_valuenotes_from_px}
\title{Clean up specific metadata from the px file}
\usage{
get_single_unit_from_px(code_no, con)

get_valuenotes_from_px(code_no, tbl_id, con)
}
\arguments{
\item{code_no}{character string of px code}

\item{tbl_id}{numeric value of table's id in the `table` table}
}
\description{
Helper functions for specific slots in the output from
\link[SURSfetchR]{get_px_metadata} that extracts and makes useable the stuff
in there
}
\keyword{internal}
