% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reading.R
\name{get_table_id}
\alias{get_table_id}
\alias{get_unit_id}
\alias{get_tab_dim_id}
\alias{get_level_value}
\alias{get_time_dimension}
\alias{get_meritve_id}
\alias{get_meritve_no}
\alias{get_level_text_from_meritve}
\alias{get_unit_levels_from_meritve}
\alias{get_valuenotes_id}
\alias{get_valuenotes_no}
\alias{get_series_id}
\alias{get_series_id_from_table}
\alias{get_last_publication_date}
\title{Helper functions to get stuff out of db tables}
\usage{
get_table_id(code_no, con)

get_unit_id(unit, con)

get_tab_dim_id(tbl_id, dim_name, con)

get_level_value(tbl_dm_id, lvl_text, con)

get_time_dimension(code_no, con)

get_meritve_id(tbl_id, con)

get_meritve_no(tbl_id, con)

get_level_text_from_meritve(meritve_dim_id, con)

get_unit_levels_from_meritve(meritve_level_text, con)

get_valuenotes_id(tbl_id, dim_name, con)

get_valuenotes_no(tbl_id, dim_name, con)

get_series_id(series_code, con)

get_series_id_from_table(tbl_id, con)

get_last_publication_date(tbl_id, con)
}
\arguments{
\item{code_no}{px. code e.g. 0300230S}

\item{con}{connection to postgres database}

\item{unit}{char text of unit label}

\item{tbl_id}{numeric table id from table table}

\item{dim_name}{character dimension name}

\item{tbl_dm_id}{numeric table-dimension id from that same table}

\item{lvl_text}{character text of level label}

\item{meritve_dim_id}{numeric dimension id of MERITVE dim}
}
\value{
numeric code from the table table

numeric code from db table unit

numeric code of table-dimension id

character code from level value code

character text of time dimension from table code

numeric id of MERITVE dimension if it exists,

numeric position of MERITVE dimension if it exists,

tibble with 4 cols including `level_value` and `unit_id`

tibble with 4 cols including `level_value` and `unit_id`

numeric id of valuenotes dimension if it exists,

numeric position of valuenotes dimension if it exists

numeric code from db table series

numeric code from db table series

numeric code from db table series
}
\description{
Group of functions that lookup stuff in relevant tables on the `con` connection
to get correct table id from px code. Also make sure the `search_path` is set to
the correct scheme.
}
\details{
NB: there is an issue with BIGINT datatype being read by dittodb in R, that's
why those have an `as.numeric()` added at the end, because there is no native
type in R or sth.
}
\keyword{internal}
