% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_category_relationship_table}
\alias{prepare_category_relationship_table}
\title{Prepare table to insert into `category_relationship` table}
\usage{
prepare_category_relationship_table(code_no, full)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{full}{full field hierarchy with parent_ids et al, output from
\link[SURSfetchR]{get_full_structure}}
}
\value{
a dataframe with the `id`, `name`, `parent_id`, `source_id` for each relationship
betweeh categories
}
\description{
Helper function that extracts the field hierarchy from the full
hierarchy data.frame, and  prepares the category relationship table with field ids and
their parents' id. Returns table ready to insert into the `category_relationship`
table with the db_writing family of functions.
}
