% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_table_table}
\alias{prepare_category_table_table}
\title{Prepare table to insert into `category_table` table}
\usage{
prepare_category_table_table(code_no, full, con, schema = "platform")
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{full}{full field hierarchy with parent_ids et al, output from}

\item{con}{connection to the database}

\item{schema}{database schema, defaults to "platform"
\link[SURSfetchR]{get_full_structure}}
}
\value{
a dataframe with the `category_id` `table_id` and `source_id` columns for
each table-category relationship.
}
\description{
Helper function that extracts the parent category for each table from the full
hierarchy data.frame, and fills up the category_table table with the table ids and
their categories (parents). A single table can have multiple parents - meaning
it is member of several categories (usually no more than two tho). Returns table
ready to insert into the `category_table`table with the db_writing family of functions.
}
