% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{prepare_data_table}
\alias{prepare_data_table}
\title{Get and prepare data for import}
\usage{
prepare_data_table(code_no, con, schema)
}
\arguments{
\item{code_no}{character SURS px. code e.g. 0300230S}

\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{schema}{character schema name, default is "platform"}
}
\value{
a dataframe the same size as was downloaded from the .px file, but
with recoded dimension levels.
}
\description{
Downloads and prepares the timeseries data for importing into the database.
Because the original data format has the full labels of the levels, these
are recoded into the alphanumeric codes.
}
