% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_dimension_levels_table}
\alias{prepare_dimension_levels_table}
\title{Prepare table to insert into `dimension_levels` table}
\usage{
prepare_dimension_levels_table(code_no, con, schema = "platform")
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{con}{connection to the database}

\item{schema}{database schema, defaults to "platform"}
}
\value{
a dataframe with the `dimension_name`, `values`, `valueTexts`, and `id`
columns for this table.
}
\description{
Helper function that extracts the levels for all the dimensions for each
table and get their codes and text.
Returns table ready to insert into the `dimension_levels`table with the
db_writing family of functions.
}
