% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_table_dimensions_table}
\alias{prepare_table_dimensions_table}
\title{Prepare table to insert into `table_dimensions` table}
\usage{
prepare_table_dimensions_table(code_no, con, schema = "platform")
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{con}{connection to the database}

\item{schema}{database schema, defaults to "platform"}
}
\value{
a dataframe with the `table_id`, `dimension_name`, `time` columns for
each dimension of this table.
}
\description{
Helper function that extracts the dimensions for each table and their "time" status.
Returns table ready to insert into the `table_dimensions`table with the
db_writing family of functions.
}
