% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_vintage_table}
\alias{prepare_vintage_table}
\title{Prepare table to insert into `vintage` table}
\usage{
prepare_vintage_table(code_no, con, schema)
}
\arguments{
\item{code_no}{character SURS px. code e.g. 0300230S}

\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{schema}{character schema name, default is "platform"}
}
\value{
a dataframe with the `series_id` and `published` columns
for all the series in this table.
}
\description{
Helper function that populates the vintage table with the new vintages. It gets
the series id's from the database and adds the publication date from the px.
}
\details{
Returns table ready to insert into the `vintage`table with the
UMARimportr::insert family of functions.
}
