% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{prepare_vintage_table}
\alias{prepare_vintage_table}
\title{Prepare table to insert into `vintage` table}
\usage{
prepare_vintage_table(code_no, con)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{con}{connection to the database}
}
\value{
a dataframe with the `series_id` and `published` columns
for all the series in this table.
}
\description{
Helper function that populated the vintage table with the new vintages. It gets
the series id's in two ways: getting all the series ids for a single table, but
also by expanding the non-time dimensions and getting all the codes for the
series and looking them up like that. This way we can catch if the dimensions
of the tables have changed in any way, and they are in fact different series. In
such a case the function doesn't return a table, but a descriptive error.
}
\details{
Returns table ready to insert into the `vintage`table with the
db_writing family of functions.
}
