% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{SURS_import_structure}
\alias{SURS_import_structure}
\title{Import structural metadata for a SURS table}
\usage{
SURS_import_structure(
  code_no,
  con,
  schema = "platform",
  all_levels = TRUE,
  keep_vintage = FALSE
)
}
\arguments{
\item{code_no}{SURS code name of the table}

\item{con}{connection to database}

\item{schema}{schema name, defaults to platform}

\item{all_levels}{logical whether to let user use dimension selector, default
is true, which automatically selects all levels in all dimensions}

\item{keep_vintage}{logical indicating whether to keep vintages, defaults to F}
}
\value{
nothing
}
\description{
Umbrella function that prepares and import all the metadata tables into
the database.
}
