% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{write_row_table_dimensions}
\alias{write_row_table_dimensions}
\title{Write categories for a single table to the `table_dimensions` table}
\usage{
write_row_table_dimensions(code_no, con, sql_statement, counter, ...)
}
\arguments{
\item{code_no}{the matrix code (e.g. 2300123S)}

\item{con}{connection to the database}

\item{sql_statement}{the sql statement to insert the values}

\item{counter}{integer counter used in  \link[SURSfetchR]{write_multiple_rows}
to count how many successful rows were inserted.}

\item{...}{just here, because other funs in this family have extra parameters
passed to them and i cannot use map unless this one also has this option.}
}
\value{
incremented counter, side effect is writing to the database.
}
\description{
Helper function that extracts the dimensions for each table and their "time" status.
Gets run from \link[SURSfetchR]{write_multiple_rows}.
}
